requirements-dev.txt
coverage==7.13.3
django-upgrade==1.30.0
djlint==1.36.4
fakeredis==2.33.0
pre-commit==4.5.1
pytest-django==4.11.1
pytest-playwright==0.7.2
ruff==0.15.0
tblib==3.2.2

-r requirements.txt